force entering the config/ subdir in ANY case
This is a general thing: when depending on binaries in subdirs AND
dependency checking for those binaries is done from within the subdir
Makefile (i.e. the decision about whether to recompile or not), we
always have to force entering the subdir or otherwise no check is
performed if the upper level dependency ("does the subdir binary
exist?") is met.