new package aufs2-util
Since the aufs2 utilities seem to be available only via git, copy the
current state into the src/ directory, with the samples/ directory
removed. The files were intentionally left untouched, so custom changes
are still kept in patch files for easier upstream sending.

Using do-extract to copy the sources to the working directory has the
downside of having to define an empty do-configure target (as this is
not needed) but allows for using update-patches just like with any
other package (which is not the case if the copying is done in a
do-configure target).