core: allow selecting parallel builds
The value of CONFIG_ADK_MAKE_JOBS is being passed to make via the option
'-j', in order to have it run multiple jobs simultaneously. This is done
for all kernel builds, and per default for each package using the
default BUILD_STYLE of 'auto', unless the package makefile defines
PKG_NOPARALLEL to a non-empty string.