change defaults for CONFIG/BUILD/INSTALL styles
All packages need an update, so here is a very huge commit.
Most of the 460 source packages use automatic style for configuration,
building and installing. Make these styles default to "auto".
If you have a package, which does not conform to this, just use
manual style and add a do-$task make target.

I added a new style named AUTOTOOL style, which is needed for some
broken packages, which needs to be updated via autoconf or automake.

I renamed CONFIGURE_STYLE to CONFIG_STYLE.

Updates for some packages, which have newer upstream versions.

Renaming of all package/*/extra directories. Use the directory
src/ to provide overwrites of source files or to add the code, when
no upstream package is available or used. src directory will be automatically
used.