libXaw: fix PKG_template invocation
For the sake of simplicity, we should stick to ${PKG_NAME} here, not the
hard-coded lower-case variant of it. This way one has to stick to the
original casing of the package name when specifying as dependency, but
that's in fact more straight forward than always having to lookup the
PKG_template invocation in question.