replace mksh scripts with faster C programs
depmaker and pkgmaker is replaced by C programs. scan-pkgs.sh will be replaced
by another mechanism. scan-pkgs.sh is needed to recognize package flavour changes,
so that a package is rebuild.

Generation of meta-data is a lot faster now.

Fix or add new PKG variables to fulfill the needs of the new programs.
Documentation will follow as soon as it is stable.