Add experimental code to self-relocate vmlinuz on e.g. brcm
This is needed for the following reason: our memory layout
looks like this:

| |vmlinux              |    |CFE| *                 |
| |                     |vmlinuz    |                |
   ^

Since CFE can only load to the spot marked with ‘^’ anyway,
we load vmlinub.elf there which is basically a big rodata
blob containing vmlinuz and minimal code moving it to the
location pointed with vmlinuz above. Another solution would
be to use CFE’s “boot -raw” to place it on the location marked
with ‘*’ above (but the CFE location and size are dynamic, and
since it insists on loading to 0x80001000 anyway, this point
is virtually moot).

Signed-off-by: Thorsten Glaser <tg@mirbsd.org>