iproute2: fix compiling tc helper libs
First of all, tc may support iptables if available. Dedicate a
sub-package for the libs generated then, so the plain tc package will
always contain the same stuff regardless of whether iptables is there or
not.
Given the right options, iproute2 is indeed able to detect iptables and
libatm, so do that instead of using a shipped Config file.
Patches updated as a nice side effect.

Signed-off-by: Phil Sutter <phil@nwl.cc>