add support for 64bit kernels on supporting targets
This is more commonly known as 32bit userland support on 64bit
architectures.

For simplicity's sake though, this implementation works the other way
round: just build a 64bit-able linker and compiler, but no
64bit-libraries at all (i.e., no multilib). This is then just enough to
compile a 64bit kernel, as that doesn't link to anything. The
alternative would have been to build a native 64bit compiler with
multilib-support in order to cross-compile a 32bit userland, resulting
in a multilib system without need for it.

In order to allow compilation of a 64bit kernel for a given target
system, have it select ADK_TARGET_KERNEL_MAY_64BIT. Upon selection of
that target, the symbol ADK_64BIT_KERNEL will occur in the "Global
settings" menu. Since certain aspects of the 64bit kernel .config may
greatly differ from it's 32bit counterpart, it has to be shipped
separately: target/<arch>/kernel64.config is the place to be.

Conflicts:

	target/Makefile
	toolchain/gcc/Makefile

Untested, due to conflicts (original patch conflicts with
multiple kernel version support).