Allow leaving /etc as untouched as possible
For customized setups, it's hard to control what files get installed
into /etc since in addition to the base-files package, any other package
may install a sample config. When two packages provide the same file,
which one of them makes it into the image depends on the order of them
being installed.

Solving this problem is not as trivial, as there are packages which must
be allowed to install stuff into /etc. Best examples are base-files and
ca-certificates.

This patch solves the problem by adding another PKG_template flag
"force_etc", which one can define in order to override the as well new
Config.in symbol "ADK_LEAVE_ETC_ALONE".