use boolean for packages
There is no need to use trstate for userland packages, because an ipkg
or txz is always build anyway. This make it more clear for the user.
A while ago a user asked me, what does it mean to use M for a package
and why it is not documented. Anyway, I think it make no sense at all.