remove dietnet patch
As discussed with phil today, patches aplied conditionally
is a bad thing and should be removed.
The dietnet patch is only available because of the mikrotik rb532
limitation of kernel size when booting from nand.

The patch is hard to maintain and will never be included upstream.
We should better use PicoTCP in such cases or develop decompression
support for this board to use compressed kernels.