package/base-files: wireless: support appending to hostapd.conf
This is useful for adding additional (virtual) access points. In my
setup, /etc/hostapd.conf.post contains:

| bss=ath6
| ssid=IPv6XS4all

In /etc/network/interfaces, there is a standard entry for ath6 which
defines the IP address.

Maybe this could be implemented in a cleaner way by introducing a
'wireless-master' setting which points to the real wireless interface
(wlan0 in my case), so ifup knows which hostapd.conf file to add the
settings to.