add new ADK Symbol PKG_NEEDS
There are a lot of packages which needs special features either
toolchain or hardware features. Add a new symbol which will be
used to disable packages, when a toolchain for example does
not provide this feature.

At the moment following features are required to set for a
package: threads rt c++.

There will follow: mmu iconv.

This will help to better support targets without MMU or
threading support.