rework prereq check
The new prereq check is completely implemented in
POSIX shell in scripts/prereq.sh.

It combines the old features from Makefile, scan-tools.sh,
scan-pkgs.sh, reloc.sh and some wrappers for tools.

The big benefit is to have all portability stuff in one place.
Furthermore we can compile GNU make and bash on the fly, for
systems lacking the required tools.
All changes on the host are detected on the fly, no make
prereq required anymore.

The build process is separated in following three phases:
1. small wrapper Makefile is used for BSD make or GNU make
2. prereq.sh is called, doing all checking, calling Makefile.adk
3. old logic in Makefile.adk or mk/build.mk is used

Tested successfully on Linux, MacOS X, Cygwin, FreeBSD, OpenBSD
and NetBSD.

An old depmaker bug was fixed, only optional host tools are compiled.
For example, even when a host provides xz, a local xz was compiled
in the past, because other packages had a build dependency on it.

Signed-off-by: Waldemar Brodkorb <wbx@openadk.org>