drop yaffs2 support
Yaffs2 is not upstream and does not work with newer 4.x Linux kernels.
We like to switch to UBIFS for any embedded devices with NAND.
Routerboards can be used with some application, which generates
a YAFFS2 image including a Linux kernel, without any YAFFS2
filesystem support in the kernel on the host or target.