rework kernel build process, changes to kernelconfig are reflected now
Cleanup the package directory which is always used as intermediate step
to build in the default some xz compressed archives with all package
data. While reconfiguring the kernel mini.config or custom config
changes to the firmware did not happen always.
Some targets where redundant and the install step tries to compile, too.

Be more quiet with any cpio usage.