remove -fhonour-copts usage
This reuires a external gcc patch, which I no longer add to newer gcc.
A lot of packages already need to disable the usage of -fhonour-copts, because
it doesn't work without patching. May be we need something like Buildroot
is using, a gcc wrapper to see poisened include or library paths while
cross-compiling.