Remove glib dependency to libelf
It looks like libelf is not used at all by glib, the way it's configured, so
remove it. The problem was that libelf would create a libelf.so.0 file that
was then mistankenly used by perf (instead of the libelf.so.1 file coming from
the elfutils package that it is really supposed to use). This is just the
minimal fix that works for us, because we are not using any other packages that
depend on libelf; to really fix the problem properly, one would have to do
what's mentioned in 2c333655f1 (get rid of libelf by porting existing users
to elfutils).